Change service provider for CentOS / RHEL 6 on upgrade from Agent 5

+++ test/kitchen/site-cookbooks/dd-agent-step-by-step/recipes/default.rb
@@ -75,7 +75,7 @@ if node['platform_family'] == 'windows'
 end
 
 service 'datadog-agent' do
-  if node['platform_family'] == 'rhel' && node['platform_version'].to_f < 7
+  if node['platform_family'] == 'rhel' && node['platform_version'].to_i < 7
     provider Chef::Provider::Service::Upstart
   end
   action :start
